Move Elem hashing stuff to its own header.
We can then set the typedef there as well...
